+++ pyoffers/models/advertiser.py
@@ -6,9 +6,7 @@ class Advertiser(Model):
     """
     An Advertiser account.
     """
-
-    def update(self, **kwargs):
-        return self._manager.update(self['id'], **kwargs)
+    generic_methods = ('update',)
 
     def block(self, reason=''):
         return self._manager.block(self['id'], reason)
@@ -17,20 +15,12 @@ class Advertiser(Model):
 class AdvertiserManager(ModelManager):
     model = Advertiser
     name = 'advertisers'
-
-    def create(self, **kwargs):
-        return self._call('create', data=kwargs)
-
-    def update(self, id, **kwargs):
-        return self._call('update', id=id, data=kwargs)
-
-    def find_by_id(self, id, contain=None):
-        return self._call('findById', id=id, contain=contain)
-
-    def find_all(self, sort=None, limit=None, page=None, contain=None, **kwargs):
-        return self._call(
-            'findAll', filters=kwargs, sort=sort, limit=limit, page=page, contain=contain, single_result=False
-        )
+    generic_methods = (
+        'create',
+        'update',
+        'find_by_id',
+        'find_all',
+    )
 
     def block(self, id, reason=''):
         return self._call('block', id=id, reason=reason)