+++ pyoffers/models/offer.py
@@ -6,10 +6,7 @@ class Offer(Model):
     """
     An Offer.
     """
-    non_model_fields = ('dne_third_party_list', )
-
-    def update(self, **kwargs):
-        return self._manager.update(self['id'], **kwargs)
+    generic_methods = ('update',)
 
     def add_target_country(self, country_code):
         return self._manager.add_target_country(self['id'], country_code)
@@ -24,20 +21,12 @@ class Offer(Model):
 class OfferManager(ModelManager):
     model = Offer
     name = 'offers'
-
-    def create(self, **kwargs):
-        return self._call('create', data=kwargs)
-
-    def update(self, id, **kwargs):
-        return self._call('update', id=id, data=kwargs)
-
-    def find_by_id(self, id, contain=None):
-        return self._call('findById', id=id, contain=contain)
-
-    def find_all(self, sort=None, limit=None, page=None, contain=None, **kwargs):
-        return self._call(
-            'findAll', filters=kwargs, sort=sort, limit=limit, page=page, contain=contain, single_result=False
-        )
+    generic_methods = (
+        'create',
+        'update',
+        'find_by_id',
+        'find_all',
+    )
 
     def add_target_country(self, id, country_code):
         return self._call('addTargetCountry', id=id, country_code=country_code)